# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/testsets/{agentId}/{checkpointId}/testset": {
      "get": {
        "operationId": "TestsetsController_getCheckpointTestset",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "description": "Pricing agent ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkpointId",
            "required": true,
            "in": "path",
            "description": "Checkpoint ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-ID",
            "required": false,
            "in": "header",
            "description": "Tenant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Testset retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointTestsetDto"
                }
              }
            }
          },
          "404": {
            "description": "Pricing agent or checkpoint not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get checkpoint testset with test data",
        "tags": [
          "testsets"
        ]
      },
      "post": {
        "operationId": "TestsetsController_aiGenerateCheckpointTestsetFromAssignedDatasets",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "description": "Pricing agent ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkpointId",
            "required": true,
            "in": "path",
            "description": "Checkpoint ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-ID",
            "required": false,
            "in": "header",
            "description": "Tenant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Testset generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointTestsetDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointTestsetDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - no assigned datasets or missing schema"
          },
          "404": {
            "description": "Pricing agent or checkpoint not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "AI generate testset from assigned datasets",
        "tags": [
          "testsets"
        ]
      }
    },
    "/testsets/{agentId}/{checkpointId}/runner": {
      "post": {
        "operationId": "TestsetsController_runCheckpointTestset",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "description": "Pricing agent ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkpointId",
            "required": true,
            "in": "path",
            "description": "Checkpoint ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-ID",
            "required": false,
            "in": "header",
            "description": "Tenant ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "failFast",
            "required": false,
            "in": "query",
            "description": "Stop execution on first unexpected failure",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tests executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointTestsetDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointTestsetDto"
                }
              }
            }
          },
          "404": {
            "description": "Pricing agent or checkpoint not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Runs checkpoint testset",
        "tags": [
          "testsets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pricing Agent Builder API",
    "description": "API documentation for the Pricing Agent Builder service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "HappyPathTestData": {
        "type": "object",
        "properties": {
          "orderInputNaturalLanguage": {
            "type": "string"
          },
          "expectedTotal": {
            "type": "number"
          },
          "expectedTotalReasoning": {
            "type": "string"
          }
        },
        "required": [
          "orderInputNaturalLanguage",
          "expectedTotal",
          "expectedTotalReasoning"
        ]
      },
...
      "UnhappyPathTestData": {
        "type": "object",
        "properties": {
          "orderInputNaturalLanguage": {
            "type": "string"
          },
          "expectedErrorType": {
            "type": "string",
            "enum": [
              "INCORRECT_INPUT_VALUE",
              "QUOTATION_RULE_VIOLATION"
            ]
          },
          "expectedErrorReasoning": {
            "type": "string"
          }
        },
        "required": [
          "orderInputNaturalLanguage",
          "expectedErrorType",
          "expectedErrorReasoning"
        ]
      },
...
      "BacktraceCalculationStepDTO": {
        "type": "object",
        "properties": {
          "operation": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "subTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BacktraceCalculationStepDTO"
            }
          }
        },
        "required": [
          "operation",
          "description"
        ]
      },
      "QuoteErrorDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "QuoteResultDTO": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "pricingCalculationBacktrace": {
            "$ref": "#/components/schemas/BacktraceCalculationStepDTO"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteErrorDTO"
            }
          }
        }
      },
...
      "TestResultDTO": {
        "type": "object",
        "properties": {
          "passed": {
            "type": "boolean"
          },
          "functionResult": {
            "$ref": "#/components/schemas/QuoteResultDTO"
          },
          "runnerException": {
            "type": "object"
          }
        },
        "required": [
          "passed"
        ]
      },
      "CheckpointHappyPathTestWithData": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "uuid"
          },
          "checkpointId": {
            "type": "string",
            "format": "uuid"
          },
          "datasetId": {
            "type": "string",
            "format": "uuid"
          },
          "datasetTestId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "functionInputParams": {
            "type": "object"
          },
          "testRunResult": {
            "$ref": "#/components/schemas/TestResultDTO"
          },
          "testData": {
            "$ref": "#/components/schemas/HappyPathTestData"
          }
        },
        "required": [
          "checkpointId",
          "datasetId",
          "datasetTestId",
          "createdAt",
          "functionInputParams",
          "testData"
        ]
      },
      "CheckpointUnhappyPathTestWithData": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "uuid"
          },
          "checkpointId": {
            "type": "string",
            "format": "uuid"
          },
          "datasetId": {
            "type": "string",
            "format": "uuid"
          },
          "datasetTestId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "functionInputParams": {
            "type": "object"
          },
          "testRunResult": {
            "$ref": "#/components/schemas/TestResultDTO"
          },
          "testData": {
            "$ref": "#/components/schemas/UnhappyPathTestData"
          }
        },
        "required": [
          "checkpointId",
          "datasetId",
          "datasetTestId",
          "createdAt",
          "functionInputParams",
          "testData"
        ]
      },
      "CheckpointTestsetDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "uuid"
          },
          "checkpointId": {
            "type": "string",
            "format": "uuid"
          },
          "happyPathTests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckpointHappyPathTestWithData"
            }
          },
          "unhappyPathTests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckpointUnhappyPathTestWithData"
            }
          }
        },
        "required": [
          "checkpointId"
        ]
      },
...
    }
  }
}